# src/openapi.yaml
openapi: '3.1.0'
info:
  title: WIZBI Control Plane API
  version: '1.0.0'
  description: |
    REST API for the WIZBI Control Plane — manages organizations, projects, 
    templates, users, and GCP/GitHub infrastructure provisioning.
    
    **Authentication**: All endpoints require either:
    - `X-Firebase-ID-Token` header (for web UI)
    - `X-API-Key` header (for programmatic/agent access)
    
    **MCP Server**: Available at `/api/mcp/sse` for AI agent integration.
  license:
    name: MIT
servers:
  - url: /api
    description: Local / deployed server
components:
  securitySchemes:
    firebaseToken:
      type: apiKey
      in: header
      name: X-Firebase-ID-Token
      description: Firebase ID Token for web UI auth
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key for programmatic access
...
security:
  - firebaseToken: []
  - apiKey: []
paths:
...
  /projects/{id}:
    get:
      tags: [Projects]
      summary: Get project details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Project details
    delete:
      tags: [Projects]
      summary: Delete project
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '202':
          description: Deletion started
...
  /users/{uid}:
    put:
      tags: [Users]
      summary: Update user roles
      parameters:
        - name: uid
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                roles:
                  type: object
                  properties:
                    superAdmin: { type: boolean }
                    orgAdmin: { type: array, items: { type: string } }
      responses:
        '200':
          description: User updated

  # ─── Settings ──────────────────────────────────────
...
  /settings/secrets/{name}:
    put:
      tags: [Settings]
      summary: Update a secret value
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value: { type: string }
      responses:
        '200':
          description: Secret updated

  # ─── API Keys ──────────────────────────────────────
...